Add 10 epochs of weights
The training script runs for 200 epochs but on my laptop it was taking
forever so I stopped after 10.  The weights are the 10 checkpoints I was
able to make while training.  The numbers in the file names correspond
to the epoch and loss respectively.

Also I am aware improvement is spelled wrong, this will be corrected in
the future.